Seperate cmd line connector interface to explicit source and sink
